Fixed imports and refactored
